Fix USB BULK buffers on 16F1459 and simplify code
The USB bulk code was not updated for 16F1459. This leaves the RX and TX
buffers outside of dual-port RAM, so it doesn't work. This commit
relocates the buffers to the right location. Instead of using absolute
addressing, the __section() specifier is used to give the compiler the
flexibility to place the buffers anywhere in the given section. This
change requires commit 611afb4 to pinguino-ide, which updates the
usbram5 section to the correct processor-specific memory ranges.

This commit also updates the CDC code to allocate those buffers in the
same way.